PR comments: Use static Socket for neli-wifi
